pyproject.toml
[project]
name = "curricerto"
version = "0.1.0"
description = "Plataforma de geração de currículos com IA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-ninja>=1.5.1",
    "django-stubs>=5.2.8",
]
